+++ src/init.php
@@ -26,6 +26,8 @@ require_once("invoice/log.php");
 require_once("dictKey/dictKey.php");
 require_once("deposit/deposit.php");
 require_once("deposit/log.php");
+require_once("brcodePayment/brcodePayment.php");
+require_once("brcodePayment/log.php");
 require_once("transfer/transfer.php");
 require_once("transfer/log.php");
 require_once("boletoPayment/boletoPayment.php");